Extract Mailer to a not Grails service class
