+++ grails-app/services/grails/plugin/awssdk/ses/AmazonSESService.groovy
@@ -1,32 +1,16 @@
 package grails.plugin.awssdk.ses
 
-import com.amazonaws.AmazonClientException
-import com.amazonaws.AmazonServiceException
-import com.amazonaws.AmazonWebServiceClient
 import com.amazonaws.auth.BasicAWSCredentials
 import com.amazonaws.regions.Region
 import com.amazonaws.regions.RegionUtils
 import com.amazonaws.services.simpleemail.AmazonSimpleEmailService
 import com.amazonaws.services.simpleemail.AmazonSimpleEmailServiceClient
-import com.amazonaws.services.simpleemail.model.*
 import grails.config.Config
 import grails.core.support.GrailsConfigurationAware
 import grails.plugin.awssdk.AwsClientUtil
 import grails.util.Environment
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
-
-import javax.activation.DataHandler
-import javax.activation.DataSource
-import javax.mail.BodyPart
-import javax.mail.Session
-import javax.mail.internet.MimeBodyPart
-import javax.mail.internet.MimeMessage
-import javax.mail.internet.MimeMultipart
-import javax.mail.util.ByteArrayDataSource
-import java.nio.ByteBuffer
-
-import static grails.plugin.awssdk.ses.AwsSdkSesEmailDeliveryStatus.*
 
 @CompileStatic
 @Slf4j
@@ -36,7 +20,7 @@ class AmazonSESService implements GrailsConfigurationAware {
     String sourceEmail
     String subjectPrefix
     String templatePath
-    AmazonWebServiceClient client
+    AwsSesMailer mailer = new AwsSesMailer()
 
     @Override
     void setConfiguration(Config co) {
@@ -62,17 +46,28 @@ class AmazonSESService implements GrailsConfigurationAware {
 
         def credentials = new BasicAWSCredentials(accessKey, secretKey)
         def clientConfiguration = AwsClientUtil.clientConfigurationWithConfig(co, defaultConfig, serviceConfig)
-        client = new AmazonSimpleEmailServiceClient(credentials, clientConfiguration)
+        mailer.client = new AmazonSimpleEmailServiceClient(credentials, clientConfiguration)
                 .withRegion(region)
+    }
+
+    private String preffixSubject(String subject) {
+        // Prefix email subject for DEV and BETA environment
+        if (Environment.current != Environment.PRODUCTION) {
+            subject = "[${Environment.current}] $subject"
+
+        } else if ( subjectPrefix ) {
+            subject = "${subjectPrefix} $subject"
+        }
+        subject
     }
 
     /**
      * @return 1 if successful, 0 if not sent, -1 if blacklisted
      */
     int mail(@DelegatesTo(TransactionalEmail) Closure composer) throws Exception {
-
-        def transactionalEmail = transactionalEmailWithClosure(composer)
-        send(transactionalEmail.destinationEmail,
+        def transactionalEmail = AwsSesMailer.transactionalEmailWithClosure(composer)
+        transactionalEmail.subject = preffixSubject(transactionalEmail.subject)
+        mailer.send(transactionalEmail.destinationEmail,
                 transactionalEmail.subject,
                 transactionalEmail.htmlBody,
                 transactionalEmail.sourceEmail,
@@ -93,122 +88,23 @@ class AmazonSESService implements GrailsConfigurationAware {
              String htmlBody,
              String sourceEmail = '',
              String replyToEmail = '') {
-        int statusId = STATUS_NOT_DELIVERED
-        if (!destinationEmail) {
-            return statusId
-        }
         if (!sourceEmail) {
             assert this.sourceEmail, "Default sourceEmail must be set in config"
             sourceEmail = this.sourceEmail
         }
-
         subject = preffixSubject(subject)
-        Destination destination = new Destination([destinationEmail])
-        Content messageSubject = new Content(subject)
-        Body messageBody = new Body().withHtml(new Content(htmlBody))
-        Message message = new Message(messageSubject, messageBody)
-        try {
-            SendEmailRequest sendEmailRequest = new SendEmailRequest(sourceEmail, destination, message)
-            if (replyToEmail) {
-                sendEmailRequest.replyToAddresses = [replyToEmail]
-            }
-            (client as AmazonSimpleEmailServiceClient).sendEmail(sendEmailRequest)
-            statusId = STATUS_DELIVERED
-        } catch (AmazonServiceException exception) {
-            if (exception.message.find("Address blacklisted")) {
-                log.debug "Address blacklisted destinationEmail=$destinationEmail"
-                statusId = STATUS_BLACKLISTED
-            } else if (exception.message.find("Missing final")) {
-                log.warn "An amazon service exception was catched while sending email: destinationEmail=$destinationEmail, sourceEmail=$sourceEmail, replyToEmail=$replyToEmail, subject=$subject"
-            } else {
-                log.warn 'An amazon service exception was catched while send +ng email' + exception.message
-            }
-        } catch (AmazonClientException exception) {
-            log.warn 'An amazon client exception was catched while sending email' + exception.message
-        }
-        statusId
+        mailer.send(destinationEmail, subject, htmlBody, sourceEmail, replyToEmail)
     }
 
-    private String preffixSubject(String subject) {
-        // Prefix email subject for DEV and BETA environment
-        if (Environment.current != Environment.PRODUCTION) {
-            subject = "[${Environment.current}] $subject"
-
-        } else if ( subjectPrefix ) {
-            subject = "${subjectPrefix} $subject"
-        }
-        subject
-    }
 
     int mailWithAttachment(@DelegatesTo(TransactionalEmail) Closure composer) throws UnsupportedAttachmentTypeException {
-        def transactionalEmail = transactionalEmailWithClosure(composer)
-        sendEmailWithAttachment(transactionalEmail)
-    }
-
-    TransactionalEmail transactionalEmailWithClosure(@DelegatesTo(TransactionalEmail) Closure composer) {
-        Closure cl = composer.clone() as Closure
-        TransactionalEmail transactionalEmail = new TransactionalEmail()
-        cl.delegate = transactionalEmail
-        cl.resolveStrategy = Closure.DELEGATE_FIRST
-        cl()
-        transactionalEmail
+        def transactionalEmail = AwsSesMailer.transactionalEmailWithClosure(composer)
+        transactionalEmail.subject = preffixSubject(transactionalEmail.subject)
+        mailer.sendEmailWithAttachment(transactionalEmail)
     }
 
     int sendEmailWithAttachment(TransactionalEmail transactionalEmail) throws UnsupportedAttachmentTypeException {
-        int statusId = STATUS_NOT_DELIVERED
-
-        Session session = Session.getInstance(new Properties())
-        MimeMessage mimeMessage = new MimeMessage(session)
-        def subject = preffixSubject(transactionalEmail.subject)
-        mimeMessage.setSubject(subject)
-        MimeMultipart mimeMultipart = new MimeMultipart()
-
-        BodyPart p = new MimeBodyPart()
-        p.setContent(transactionalEmail.htmlBody, "text/html")
-        mimeMultipart.addBodyPart(p)
-
-        for(TransactionalEmailAttachment attachment : transactionalEmail.attachments) {
-
-            if(!AwsSdkSesMimeType.isMimeTypeSupported(attachment.mimeType)) {
-                throw new UnsupportedAttachmentTypeException()
-            }
-
-            MimeBodyPart mimeBodyPart = new MimeBodyPart()
-            mimeBodyPart.setFileName(attachment.filename)
-            mimeBodyPart.setDescription(attachment.description, "UTF-8")
-            DataSource ds = new ByteArrayDataSource(new FileInputStream(new File(attachment.filepath)), attachment.mimeType)
-            mimeBodyPart.setDataHandler(new DataHandler(ds))
-            mimeMultipart.addBodyPart(mimeBodyPart);
-        }
-        mimeMessage.content = mimeMultipart
-
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream()
-        mimeMessage.writeTo(outputStream)
-        RawMessage rawMessage = new RawMessage(ByteBuffer.wrap(outputStream.toByteArray()))
-
-        SendRawEmailRequest rawEmailRequest = new SendRawEmailRequest(rawMessage)
-
-        rawEmailRequest.setDestinations(transactionalEmail.recipients)
-        rawEmailRequest.setSource(transactionalEmail.sourceEmail)
-
-        try {
-            (client as AmazonSimpleEmailServiceClient).sendRawEmail(rawEmailRequest);
-            statusId = STATUS_DELIVERED
-
-        } catch (AmazonServiceException exception) {
-            if (exception.message.find("Address blacklisted")) {
-
-                log.debug "Address blacklisted destinationEmail=${transactionalEmail.recipients.toString()}"
-                statusId = STATUS_BLACKLISTED
-            } else if (exception.message.find("Missing final")) {
-                log.warn "Invalid parameter value: destinationEmail=${transactionalEmail.recipients.toString()}, sourceEmail=${transactionalEmail.sourceEmail}, replyToEmail=${transactionalEmail.replyToEmail}, subject=${subject}"
-            } else {
-                log.warn 'An amazon service exception was catched while sending email with attachment' + exception.message
-            }
-        } catch (AmazonClientException exception) {
-            log.warn 'An amazon client exception was catched while sending email with attachment' + exception.message
-
-        }
-        statusId
+        transactionalEmail.subject = preffixSubject(transactionalEmail.subject)
+        mailer.sendEmailWithAttachment(transactionalEmail)
     }
 }
