+++ src/test/groovy/grails/plugin/awssdk/ses/AmazonSESServiceSpec.groovy
@@ -17,20 +17,20 @@ class AmazonSESServiceSpec extends Specification {
 
     void setup() {
         // Mock collaborator
-        service.client = Mock(AmazonSimpleEmailServiceClient)
+        service.mailer.client = Mock(AmazonSimpleEmailServiceClient)
     }
 
     void setUpSesInteractions(Map parameters = [:]) {
         if (parameters.serviceException) {
-            1 * service.client.sendEmail(_) >> {
+            1 * service.mailer.client.sendEmail(_) >> {
                 throw new AmazonServiceException(parameters.serviceException)
             }
         } else if (parameters.clientException) {
-            1 * service.client.sendEmail(_) >> {
+            1 * service.mailer.client.sendEmail(_) >> {
                 throw new AmazonClientException(parameters.clientException)
             }
         } else {
-            1 * service.client.sendEmail(_) >> [messageId: 'someMessageId']
+            1 * service.mailer.client.sendEmail(_) >> [messageId: 'someMessageId']
         }
     }
 
@@ -115,67 +115,5 @@ class AmazonSESServiceSpec extends Specification {
     }
 
 
-    void "test transactionalEmailWithClosure"() {
-        when:
-        TransactionalEmail transactionalEmail = service.transactionalEmailWithClosure {
-            subject 'Hi Paul'
-            htmlBody '<p>This is an example body</p>'
-            to 'me@sergiodelamo.com'
-            from 'subscribe@groovycalamari.com'
-            attachment {
-                filename 'test.pdf'
-                filepath '/tmp/test.pdf'
-                mimeType 'application/pdf'
-                description 'An example pdf'
-            }
-        }
-
-        then:
-        transactionalEmail
-        transactionalEmail.subject == 'Hi Paul'
-        transactionalEmail.htmlBody == '<p>This is an example body</p>'
-        transactionalEmail.sourceEmail == 'subscribe@groovycalamari.com'
-        transactionalEmail.recipients == ['me@sergiodelamo.com']
-        transactionalEmail.destinationEmail == 'me@sergiodelamo.com'
-        transactionalEmail.attachments.size() == 1
-        transactionalEmail.attachments.first().filename == 'test.pdf'
-        transactionalEmail.attachments.first().filepath == '/tmp/test.pdf'
-        transactionalEmail.attachments.first().mimeType == 'application/pdf'
-        transactionalEmail.attachments.first().description == 'An example pdf'
-
-
-        when:
-        def f = new File('src/integration-test/groovy/grails/plugin/awssdk/ses/groovylogo.png')
-
-        then:
-        f.exists()
-
-
-        when:
-        transactionalEmail = service.transactionalEmailWithClosure {
-            subject 'Hi Paul'
-            htmlBody '<p>This is an example body</p>'
-            to 'me@sergiodelamo.com'
-            from 'subscribe@groovycalamari.com'
-            attachment {
-                filepath f.absolutePath
-            }
-        }
-
-        then:
-        transactionalEmail
-        transactionalEmail.subject == 'Hi Paul'
-        transactionalEmail.htmlBody == '<p>This is an example body</p>'
-        transactionalEmail.sourceEmail == 'subscribe@groovycalamari.com'
-        transactionalEmail.recipients == ['me@sergiodelamo.com']
-        transactionalEmail.destinationEmail == 'me@sergiodelamo.com'
-        transactionalEmail.attachments.size() == 1
-        transactionalEmail.attachments.first().filename == 'groovylogo.png'
-        transactionalEmail.attachments.first().filepath == f.absolutePath
-        transactionalEmail.attachments.first().mimeType == 'image/png'
-        transactionalEmail.attachments.first().description == ''
-
-
-    }
 
 }
